Create public/index.html for video player
This commit adds the main index.html file to the public folder. It contains the basic structure of a web page with a video player, progress bar, control buttons, and a search popup for movies. The scripts dependencies such as "socket.io", "jquery", and "interface.js" are also included at the bottom of the body.